Inner structure of servlets has been simplified. MathServlet abstract class has been removed.
