+++ src/com/test/servlet/math/DivideServlet.java
@@ -2,22 +2,45 @@ package com.test.servlet.math;
 
 import java.io.IOException;
 
+import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/div")
-public class DivideServlet extends MathServlet {
-	
-	private static final long serialVersionUID = -7788488297965784175L;
+import com.test.servlet.util.Validator;
 
-	@Override
-	protected void mathOperation(int a, int b, HttpServletResponse response) throws IOException {
+@WebServlet("/div")
+public class DivideServlet extends HttpServlet{
+
+	private static final long serialVersionUID = 485135717800530684L;
+
+	public DivideServlet() {
+		super();
+	}
+
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+
+		String firstInputParameter = request.getParameter("a");
+		String secondInputParameter = request.getParameter("b");
+
+		Validator.validate(firstInputParameter, secondInputParameter, response);
+
+		int a = Integer.valueOf(firstInputParameter);
+		int b = Integer.valueOf(secondInputParameter);
 		
-		try{
-		response.getWriter().append("Division is: " + (a / b));}
-		catch(ArithmeticException ae){
-			response.getWriter().append("Please input non-zero divider!");
+		if(b == 0){
+			response.getWriter().append("Please enter non-zero divider!");
+			response.flushBuffer();
 		}
+
+		response.getWriter().append("Division is: " + (a / b));
+	}
+
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		doGet(request, response);
 	}
 
 }
